backend/requirements.txt
fastapi
uvicorn
python-telegram-bot
requests
apscheduler
python-dotenv
crewai
crewai-tools
google-api-python-client
chromadb
SQLAlchemy
pydantic
pydantic-settings
typing-extensions
google-genai
multiprocess
pytz